Backend/requirements.txt
attrs==25.4.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mistune==3.1.4
packaging==25.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
Flask-JWT-Extended==4.6.0
bcrypt==4.2.0
